Support gdn fused_rms_norm_gated
